+++ uefi-macros/tests/ui/fail/entry_bad_arg.rs
@@ -1,11 +1,8 @@
-#![allow(unused_imports)]
 #![no_main]
-#![allow(deprecated)]
 
 use uefi::prelude::*;
-use uefi_macros::entry;
 
 #[entry]
-fn main(_handle: Handle, _st: SystemTable<Boot>, _x: usize) -> Status {
+fn main(_x: usize) -> Status {
     Status::SUCCESS
 }